Promote mock_config_entry_with_heat_request fixture to conftest
Removes duplicate fixture definitions from test_valve_sync.py and
test_entity_unavailability.py.
